apps/pipelines/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "riskops-pipelines"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "typer==0.12.3",
  "click==8.1.7",
  "pandas==2.2.3",
  "numpy==2.2.1",
  "psycopg2-binary==2.9.9",
  "sqlalchemy==2.0.46",
  "mlflow==2.19.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["riskops_pipelines*"]

